Fixes file uploads with AJAX.
Because of the way we were checking the URL to conditionally run the form
alter, our submission handler was not being added after an AJAX upload
occured. This was most obvious when doing multi file uploads in which, the
only way to add a second item (when JS is enabled) is to upload a file which
triggers a form re-render. This stashes data in the form_state which can be
checked as part of the logic to run the form alter.